Add support for OpenSuse 13.1 support.
Added defaults based on a fresh OpenSuse 13.1 installation to
init.pp. template got updated, to be able to manage other
entries that were so far static in the file.
Other OS/Distributions got updated to have the defaults
as it was in the static template.

Also mentioning it in README.md and metadata.json, and add
spec test for it.